container/requirements.txt
flask
pytest
selenium
pytest-html
requests
gunicorn
google-cloud-storage
